backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
branca==0.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.14
folium==0.20.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.2.6
openai==1.92.3
openpyxl==3.1.5
pandas==2.3.0
passlib==1.7.4
psycopg2-binary==2.9.10
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2025.2
pywin32==310
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
xyzservices==2025.4.0